# docs/swagger.yaml
definitions:
  Middlewares.Response:
    properties:
      data:
        type: object
      errmsg:
        type: string
      errno:
        type: integer
    type: object
  Models.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
host: localhost:8088
info:
  contact:
    email: 790227542@qq.com
    name: www.topgoer.com
    url: https://www.topgoer.com
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://www.topgoer.com
  title: 自定义项目API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录接口
      parameters:
      - description: 登录参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/Models.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Middlewares.Response'
      security:
      - ApiKeyAuth: []
      summary: 用户登录
      tags:
      - 鉴权相关接口
  /auth/logout:
    get:
      consumes:
      - application/json
      description: 注销登录接口
      parameters:
      - description: 用户令牌
        in: header
        name: token
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Middlewares.Response'
      security:
      - ApiKeyAuth: []
      summary: 注销登录
      tags:
      - 鉴权相关接口
...
swagger: "2.0"